score calculation

+++ ZuciQuizLibrary/Models/Score.cs
@@ -9,11 +9,11 @@ using System.Threading.Tasks;
 namespace ZuciQuizLibrary.Models
 {
     [Table("Score")]
-    public class Score : CommonEntity  //  score
+    public class Score : CommonEntity  
     {
         public int UserId { get; set; } 
         public int TopicId { get; set; } 
-        public int Mark { get; set; }
+        public Double Mark { get; set; }
         public DateTime DateCompleted { get; set; }
         public User? User { get; set; }
         public Topic? Topic { get; set; }
